Update 2021-03-26-freedom-of-speech.md
Removed the image as it was doing something weird...The image was initially visible when I declared it in the `front matter` but it now appears to be missing after I tried adding an alternate title with `alt`. I tried moving the image inside the file but it is only visible if I switch to read mode on the browser.